refactor: clean up stale main.py and examples folder
- Extract MCPClient to tools/mcp/core/client.py for proper reuse
- Update Gaea2 tests to import from new location
- Remove stale main.py from root (replaced by MCP server architecture)
- Remove outdated examples folder (superseded by test scripts)
- Update documentation to reflect new structure

This cleanup removes artifacts from the early architecture that have
been superseded by the modular MCP server design. Each MCP server
now has its own test script that demonstrates proper usage.
