+++ tests/gaea2/test_gaea2_templates.py
@@ -12,7 +12,7 @@ from typing import Any, Dict, Tuple
 # Add project root to path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 
-from main import MCPClient  # noqa: E402
+from tools.mcp.core import MCPClient  # noqa: E402
 
 
 class Gaea2TemplateTester:
