+++ tools/mcp/core/__init__.py
@@ -1,7 +1,8 @@
 """Core utilities and base classes for MCP servers"""
 
 from .base_server import BaseMCPServer
+from .client import MCPClient
 from .http_bridge import HTTPBridge
 from .utils import setup_logging, validate_environment
 
-__all__ = ["BaseMCPServer", "HTTPBridge", "setup_logging", "validate_environment"]
+__all__ = ["BaseMCPServer", "MCPClient", "HTTPBridge", "setup_logging", "validate_environment"]
